STK: Add content information to the icon that is displayed.
Add an image view to the layout of the activities( if not already present)
which displays the Stk activity's icon. An appropriate content
description is added to the icon's image view . The icon's content
information helps in testing various icon related stk test cases using the
UI Automator.

Support has been added for following commands:
1. Display Text
2. Get Inkey
3. Get Input
4. Set Up Menu
5. Send Short Message
6. Send SS
7. Send USSD
8. Set Up Call
9. Send DTMF

Change-Id: Ib9eeea7169f1f66bbb9bc1ec3081927ae998b3a2
CRs-Fixed: 823505